Configuration/requirements.txt
# Core dependencies
openai>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# PDF processing (for documents in data/raw/)
PyPDF2>=3.0.0
pypdf>=3.0.0

# Web interface
streamlit>=1.28.0

# Environment management
python-dotenv>=1.0.0

# Online hosting
pyngrok>=7.0.0

# Optional: For better text processing
nltk>=3.8
spacy>=3.6.0

# Optional: For advanced document processing
langchain>=0.1.0
tiktoken>=0.5.0